+++ install.py
@@ -1,4 +1,0 @@
-#!/usr/bin/python3
-import psml
-if __name__=="__main__":
-    psml.__install__()